Add files via upload
In these pictures you can see the preliminary analysis i performed on a portion of the well. the first step was to identify places of change in mud, or gain and loss in the wellbore to be more precise. To do so, I decided to find the second derivative of the mud level/pit volume data set, to determine places of quick changes, then I'd try to correlate those to the change in mu volume. 
Next, i tried ti find a pattern with the back pressure, which will determine if indeed something happened that required the back pressure to kick in. Last, I checked with MSE how the plot responded. 
Blue is mud volume, and red is pressure, black is MSE.